refactor: move INotificationService to Common/Interfaces
Moved INotificationService to the correct application interface layer according to the Clean Architecture standards of the project. Updated all corresponding namespaces in endpoints and dependency injection configurations.
