+++ backend/Valora.Api/Endpoints/NotificationEndpoints.cs
@@ -1,3 +1,4 @@
+using Valora.Application.Common.Interfaces;
 using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
 using Valora.Application.Services;
